Convert raw item searching to TS
